Fix : Move dtos & enums package to model package
