+++ src/main/java/kr/co/smkpetclinicstudy/controller/OwnerController.java
@@ -1,8 +1,8 @@
 package kr.co.smkpetclinicstudy.controller;
 
 import jakarta.validation.Valid;
-import kr.co.smkpetclinicstudy.service.model.request.OwnerReqDTO;
-import kr.co.smkpetclinicstudy.service.model.response.OwnerResDTO;
+import kr.co.smkpetclinicstudy.service.model.dtos.request.OwnerReqDTO;
+import kr.co.smkpetclinicstudy.service.model.dtos.response.OwnerResDTO;
 import kr.co.smkpetclinicstudy.service.service.OwnerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
