+++ src/main/java/kr/co/smkpetclinicstudy/service/service/PetService.java
@@ -4,8 +4,8 @@ import kr.co.smkpetclinicstudy.persistence.entity.Owner;
 import kr.co.smkpetclinicstudy.persistence.entity.Pet;
 import kr.co.smkpetclinicstudy.persistence.repository.OwnerRepository;
 import kr.co.smkpetclinicstudy.persistence.repository.PetRepository;
-import kr.co.smkpetclinicstudy.service.model.request.PetReqDTO;
-import kr.co.smkpetclinicstudy.service.model.response.PetResDTO;
+import kr.co.smkpetclinicstudy.service.model.dtos.request.PetReqDTO;
+import kr.co.smkpetclinicstudy.service.model.dtos.response.PetResDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
