+++ src/main/java/kr/co/smkpetclinicstudy/service/service/VetService.java
@@ -5,8 +5,8 @@ import kr.co.smkpetclinicstudy.persistence.entity.Vet;
 import kr.co.smkpetclinicstudy.persistence.entity.VetSpecialty;
 import kr.co.smkpetclinicstudy.persistence.repository.SpecialtyRepository;
 import kr.co.smkpetclinicstudy.persistence.repository.VetRepository;
-import kr.co.smkpetclinicstudy.service.model.request.VetReqDTO;
-import kr.co.smkpetclinicstudy.service.model.response.VetResDTO;
+import kr.co.smkpetclinicstudy.service.model.dtos.request.VetReqDTO;
+import kr.co.smkpetclinicstudy.service.model.dtos.response.VetResDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
